refactor: Move MessageFlag(s) to its own file
