+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Dev: Refactor/unsingletonize `UserDataController`. (#5459)
 - Dev: Cleanup `BrowserExtension`. (#5465)
 - Dev: Deprecate Qt 5.12. (#5396)
+- Dev: Refactored `MessageFlag` into its own file. (#5549)
 - Dev: The running Qt version is now shown in the about page if it differs from the compiled version. (#5501)
 - Dev: `FlagsEnum` is now `constexpr`. (#5510)
 - Dev: Documented and added tests to RTL handling. (#5473)
