+++ src/messages/Message.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
-#include "common/FlagsEnum.hpp"
+#include "messages/MessageFlag.hpp"
 #include "providers/twitch/ChannelPointReward.hpp"
 #include "util/QStringHash.hpp"
 
-#include <magic_enum/magic_enum.hpp>
 #include <QColor>
 #include <QTime>
 
@@ -18,53 +17,6 @@ class MessageElement;
 class MessageThread;
 class Badge;
 class ScrollbarHighlight;
-
-enum class MessageFlag : int64_t {
-    None = 0LL,
-    System = (1LL << 0),
-    Timeout = (1LL << 1),
-    Highlighted = (1LL << 2),
-    DoNotTriggerNotification = (1LL << 3),  // disable notification sound
-    Centered = (1LL << 4),
-    Disabled = (1LL << 5),
-    DisableCompactEmotes = (1LL << 6),
-    Collapsed = (1LL << 7),
-    ConnectedMessage = (1LL << 8),
-    DisconnectedMessage = (1LL << 9),
-    Untimeout = (1LL << 10),
-    PubSub = (1LL << 11),
-    Subscription = (1LL << 12),
-    DoNotLog = (1LL << 13),
-    AutoMod = (1LL << 14),
-    RecentMessage = (1LL << 15),
-    Whisper = (1LL << 16),
-    HighlightedWhisper = (1LL << 17),
-    Debug = (1LL << 18),
-    Similar = (1LL << 19),
-    RedeemedHighlight = (1LL << 20),
-    RedeemedChannelPointReward = (1LL << 21),
-    ShowInMentions = (1LL << 22),
-    FirstMessage = (1LL << 23),
-    ReplyMessage = (1LL << 24),
-    ElevatedMessage = (1LL << 25),
-    SubscribedThread = (1LL << 26),
-    CheerMessage = (1LL << 27),
-    LiveUpdatesAdd = (1LL << 28),
-    LiveUpdatesRemove = (1LL << 29),
-    LiveUpdatesUpdate = (1LL << 30),
-    /// The header of a message caught by AutoMod containing allow/disallow
-    AutoModOffendingMessageHeader = (1LL << 31),
-    /// The message caught by AutoMod containing the user who sent the message & its contents
-    AutoModOffendingMessage = (1LL << 32),
-    LowTrustUsers = (1LL << 33),
-    /// The message is sent by a user marked as restricted with Twitch's "Low Trust"/"Suspicious User" feature
-    RestrictedMessage = (1LL << 34),
-    /// The message is sent by a user marked as monitor with Twitch's "Low Trust"/"Suspicious User" feature
-    MonitoredMessage = (1LL << 35),
-    /// The message is an ACTION message (/me)
-    Action = (1LL << 36),
-};
-using MessageFlags = FlagsEnum<MessageFlag>;
 
 struct Message;
 using MessagePtr = std::shared_ptr<const Message>;
@@ -113,8 +65,3 @@ struct Message {
 };
 
 }  // namespace chatterino
-
-template <>
-struct magic_enum::customize::enum_range<chatterino::MessageFlag> {
-    static constexpr bool is_flags = true;
-};
